@@ -12,30 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "@angular/service-worker": "^5.2.5",
+    "@angular-devkit/schematics": "^0.7.0-beta.1",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@angular/service-worker": "^6.0.3",
+    "@ngxs/store": "^3.1.3",
     "@types/youtube": "0.0.32",
     "core-js": "^2.5.3",
-    "ngx-youtube-player": "^0.1.0",
-    "rxjs": "^5.5.6",
+    "ngx-youtube-player": "^6.0.0",
+    "rxjs": "^6.2.0",
     "youtube": "^0.1.0",
     "youtube-player": "^5.5.0",
-    "zone.js": "^0.8.20"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.5",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -51,6 +53,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
   }
 }
