@@ -1,9 +1,9 @@
 import { NgModule }                           from '@angular/core';
 import { NoopAnimationsModule }                 from '@angular/platform-browser/animations';
-import { MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule } from '@angular/material';
+import { MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule, MatTooltipModule } from '@angular/material';
 
 @NgModule({
-  imports: [ NoopAnimationsModule, MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule ],
-  exports: [ NoopAnimationsModule, MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule ],
+  imports: [ NoopAnimationsModule, MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule, MatTooltipModule ],
+  exports: [ NoopAnimationsModule, MatCommonModule, MatButtonModule, MatSidenavModule, MatListModule, MatExpansionModule, MatCardModule, MatTooltipModule ],
 })
 export class AppMaterialsModule { }
